Create AnalyzerBase
https://github.com/Not-Endurance/not-timing-system/issues/293
